requirements.txt
PyYAML==6.0.2
pyserial==3.5
python-dotenv==1.0.1
aiogram==3.17.0
asyncpg==0.30.0
APScheduler==3.11.0
matplotlib==3.10.0
